[library-explorer] Separate ShelfProgressBar from ClassSlider
